fix(api-converse): serve /chat/api-converse from app-api, not via inference proxy
The API-key converse endpoint was broken in cloud. app-api proxied
POST /chat/api-converse to `{INFERENCE_API_URL}/chat/api-converse`, but
inference-api now runs inside an AgentCore Runtime whose data plane only
serves POST /invocations and GET /ping — any other path returns
UnknownOperationException (404) before reaching the container. It worked
locally only because localhost:8001 bypasses the runtime gateway.

Relocate the handler onto app-api as a self-contained route (validate key
-> RBAC -> bedrock-runtime.converse -> cost accounting), reusing the shared
services it already depends on. app-api reaches Bedrock directly via its
task role, so there is no inference-api hop and no INFERENCE_API_URL
dependency. Delete the proxy, the now-dead inference-api route, and its
DTOs (moved to app_api/chat/models.py). Repoint the converse tests at the
app-api module.

Verified: 263 backend tests pass, import-boundary test clean, and a real
un-mocked smoke against a Bedrock model returns 200 (stream + non-stream).
